fix: 반복되는 코드 더 간략하게 TaskItem과 Task로 분리
