+++ src/components/styles/Task.styles.js
@@ -1,0 +1,16 @@
+import styled from "@emotion/styled";
+
+export const Style = {
+  Wrapper: styled.div`
+    width: 100%;
+    height: 200px;
+    overflow-y: scroll;
+    margin-top: 10px;
+  `,
+  ItemWrapper: styled.div`
+    display: flex;
+    align-items: center;
+    width: 90%;
+    height: 40px;
+  `,
+};
